Create random obj.java
This project demonstrates the use of a Random object to generate a random number and a Scanner object to read user input. The program generates a random number between 1 and 3, which corresponds to rock, paper, or scissors. The user is prompted to enter their choice, and the program displays the computer's choice and the result of the game.